Use BigInteger instead of Integer for storing discord ids

+++ bot/tables.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, PrimaryKeyConstraint, Integer
+from sqlalchemy import Column, PrimaryKeyConstraint, Integer, BigInteger
 from sqlalchemy.orm import declarative_base
 
 Base = declarative_base()
@@ -6,7 +6,7 @@ Base = declarative_base()
 class UserXP(Base):
 	__tablename__ = 'userxp'
 
-	discord_id = Column(Integer, primary_key=True)
+	discord_id = Column(BigInteger, primary_key=True)
 	xp = Column(Integer, nullable=False)
 
 	def __repr__(self):
